add active-sequences trace replay test
