+++ lib/bench/Cargo.toml
@@ -56,5 +56,6 @@ minstant = "0.1.7"
 plotters = { version = "0.3", default-features = false, features = ["svg_backend", "line_series", "point_series", "full_palette"] }
 tokio = { workspace = true, features = ["rt", "macros", "time"] }
 tokio-util = { workspace = true }
+tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 uuid = { workspace = true }
